requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
websockets>=12.0
argon2-cffi>=23.1.0
pydantic>=2.0.0
